make utils module private and provide a public utils module with shortcuts
